+++ packages/pwa-kit-dev/src/configs/webpack/loaders/extensions-loader.js
@@ -73,6 +73,8 @@ module.exports = function () {
                     .join(',\n                ')}
             ]
 
+            // TODO: Once we create @salesforce/pwa-kit-extensibility we will refactor this code to use a shared utility that
+            // normalizes/expands the configuration array before it is uses in this loader.
             const normalizeExtensionsList = (extensions = []) =>
                 extensions.map((extension) => {
                     return {